feat: Platform specific design changes on pagination component
Split pagination into a base and platform specific components.

REPLAT-1297
