+++ .gitignore
@@ -67,3 +67,5 @@ fructose/components.js
 fructose/components.test.js
 
 dextrose/snappy/**
+
+.vscode/**
