+++ packages/author-profile/author-profile-content.js
@@ -108,12 +108,11 @@ class AuthorProfileContent extends React.Component {
     };
 
     const paginationComponent = (
-      { hideResults = false, hideTopKeyline = false } = {}
+      { hideResults = false} = {}
     ) => (
       <AuthorProfilePagination
         count={count}
         hideResults={hideResults}
-        hideTopKeyline={hideTopKeyline}
         onNext={(...args) => {
           onNext(...args);
           scrollToTopNextFrame();
@@ -185,7 +184,7 @@ class AuthorProfileContent extends React.Component {
               twitter={twitter}
               onTwitterLinkPress={onTwitterLinkPress}
             />
-            {paginationComponent({ hideResults: false, hideTopKeyline: true })}
+            {paginationComponent({ hideResults: false})}
           </View>
         }
         ListFooterComponent={paginationComponent({ hideResults: true })}
